Chose to take in consideration client time zone on the server, the application being developed for French users. Another solution might be thought about later.
Added a DateTool on the server which should be used every time we have to deal with dates.

Search on meals by date changed to now look at meals from the start of the current day to the start of the day after, speaking in "European/Paris" time.

Moved server tools in a specific package.