Resolved bug:
Fields even if they are not annotated with "@Persistent" seem to be persisted along with the Object owning them.

For instance, Order contains three Meals and a User. Meals and User were replicated on the datastore each time the Order was saved (persisted).

The solution adopted is to assign null to these fields before persisting the Order.